+++ src/Components/Layout/Input/Root.tsx
@@ -16,6 +16,10 @@ export default function InputRoot(props: {
 }) {
 
   const { theme } = useMemo(() => ConfigService.config, []);
+
+  const noLabel = props.label !== undefined && props.label !== '';
+  const noIconButtons = props.iconButtons === undefined;
+
   const backgroundColor = props.backgroundColor ?? theme.background;
   const color = props.color ?? theme.onBackground;
 
@@ -23,42 +27,44 @@ export default function InputRoot(props: {
     <View
       style={{
         paddingHorizontal: 5,
-        paddingTop: 15,
+        paddingTop: (noLabel && noIconButtons) ? 5 : 15,
       }}
     >
-      {(props.label !== undefined && props.label !== '') && (<RootText
-        style={{
-          position: 'absolute',
-          backgroundColor: backgroundColor,
-          color: color,
-          fontSize: 20,
-          paddingHorizontal: 5,
-          top: 0,
-          left: 15,
-          zIndex: 1,
-        }}
-      >
-        {props.label}
-      </RootText>)}
-      <View
-        style={{
-          position: 'absolute',
-          flexDirection: 'row',
-          backgroundColor: backgroundColor,
-          zIndex: 1,
-          height: 30,
-          top: 0,
-          right: 15,
-        }}
-      >
-        {props.iconButtons}
-      </View>
+      {(props.label !== undefined && props.label !== '') && (
+        <RootText
+          style={{
+            position: 'absolute',
+            backgroundColor: backgroundColor,
+            color: color,
+            fontSize: 20,
+            paddingHorizontal: 5,
+            top: 0,
+            left: 15,
+            zIndex: 1,
+          }}
+        >
+          {props.label}
+        </RootText>
+      )}
+      {!noIconButtons && (
+        <View
+          style={{
+            position: 'absolute',
+            flexDirection: 'row',
+            backgroundColor: backgroundColor,
+            zIndex: 1,
+            height: 30,
+            top: 0,
+            right: 15,
+          }}
+        >
+          {props.iconButtons}
+        </View>
+      )}
       <View
         style={[{
           width: '100%',
           paddingHorizontal: 10,
-          paddingTop: 20,
-          paddingBottom: 10,
           gap: 10,
           backgroundColor: backgroundColor,
           borderColor: theme.primary,